Refactor out the server code back into the server
